ignore "None" Factors in looking for generic factors
The search for generic_factors method was failing by searching None for generic factors. This is suspicious because if it's the correct solution, I don't see why it wasn't needed earlier. The problem originated either when I separated Holding objects from Rule object, or when I gave Holding objects the ability to call Rule.__init__