make get_references_from_mentioned a schema method
This was an attempt to get one nested schema to add entries to "mentioned" and have the parent schema or other nested schemas access it. That doesn't seem to be possible. The new plan is to use the schemas mostly for dumping and have their load methods call the corresponding functions in the readers module.