assign names to factors in process of collecting name index
Previously the names were collected in the process of expanding shorthand in
raw files to be loaded as Factors. The old method seems not to work because
the names of some factors need to be included in the names of others, and
retrieving names to incorporate them into other names requires them to have
been saved to the name index.