collect_anchors_from_index filters out pipe symbols
Still unclear where the pipe symbol is coming from.

Filtering it out is a bad alternative to preventing it from
being created in the first place.

Preventing the pipe symbol from being at the start of a string
in the YAML file was not enough to prevent it from appearing
in the "anchors" field and being passed to split_anchor_text.